requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.0.4
django-crispy-forms==2.1
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
sqlparse==0.5.0
whitenoise==6.6.0
